Fail fast in the install script so that 'success' isn't echo'd on failure.

+++ install.sh
@@ -1,24 +1,33 @@
 #!/bin/bash
 source ./variables.sh
 
-SCREENSHOT_DIR_TO_USE=$(defaults read com.apple.screencapture location)
+# Fail fast for any commands which return an error
+set -e
 
 AUTOMATOR_IMGUR_SCREENSHOT_UPLOAD_TEMPLATE_DIR=./automator/imgur-screenshot-upload-template.workflow
 AUTOMATOR_IMGUR_UPLOAD_TEMPLATE_DIR=./automator/imgur-upload-template.workflow
-
 AUTOMATOR_DOCUMENT_FILE=/Contents/document.wflow
 
+# Determine the directory in which screenshots will be saved
 if [ "x$1" != "x" ]; then
     SCREENSHOT_DIR_TO_USE=$1
-    ./set_screenshot_dir.sh $SCREENSHOT_DIR_TO_USE &> /dev/null
+	./set_screenshot_dir.sh $SCREENSHOT_DIR_TO_USE
+elif ! SCREENSHOT_DIR_TO_USE=$(defaults read com.apple.screencapture location); then
+	# Screenshot dir is OS default by omission of any preference to the contrary
+    echo "screenshot dir default: \"$HOME/Desktop\""
 fi
 
-echo "copying imguru to $IMGURU_INSTALL_PATH"
+if [ -z ${SCREENSHOT_DIR_TO_USE+x} ]; then
+    echo "screenshot dir is: \"$SCREENSHOT_DIR_TO_USE\""
+fi
 
-# copy the binary to upload images
-cp ./imgur-uploader/imguru $IMGURU_INSTALL_PATH
-
-echo "screenshot dir is: \"$SCREENSHOT_DIR_TO_USE\""
+# copy the binary to upload images if it's not already on the PATH
+if IMGURU_EXISTING_PATH=$(which imguru); then
+	echo "existing imguru found $IMGURU_EXISTING_PATH"
+else
+    echo "copying imguru to $IMGURU_INSTALL_PATH"
+    cp ./imgur-uploader/imguru $IMGURU_INSTALL_PATH
+fi
 
 echo "installing..."
